+ before switching to inertia recording

+++ BOLDApp/src/au/edu/melbuni/boldapp/Recognizer.java
@@ -2,7 +2,7 @@ package au.edu.melbuni.boldapp;
 
 public abstract class Recognizer {
 
-	public abstract boolean isSilence(byte[] buffer);
+	public abstract boolean isSilence(short[] buffer);
 
 	public abstract boolean isSpeech(short[] buffer);
