+++ BOLDApp/src/au/edu/melbuni/boldapp/SpeechTriggers.java
@@ -2,8 +2,8 @@ package au.edu.melbuni.boldapp;
 
 public interface SpeechTriggers {
 
-	public void silenceTriggered(byte[] buffer, boolean justChanged);
-	public void speechTriggered(byte[] buffer, boolean justChanged);
-	public void neitherTriggered(byte[] buffer);
+	public void silenceTriggered(short[] buffer, boolean justChanged);
+	public void speechTriggered(short[] buffer, boolean justChanged);
+	public void neitherTriggered(short[] buffer);
 	
 }
